Conflicts:
	extras/Introjucer/Source/Project Saving/jucer_ProjectExport_XCode.h
	modules/juce_audio_plugin_client/AAX/juce_AAX_Wrapper.cpp
	modules/juce_audio_plugin_client/VST3/juce_VST3_Wrapper.cpp
	modules/juce_audio_processors/format_types/juce_AudioUnitPluginFormat.mm

+++ modules/juce_audio_formats/codecs/juce_LAMEEncoderAudioFormat.cpp
@@ -30,9 +30,9 @@ public:
     Writer (OutputStream* destStream, const String& formatName,
             const File& appFile, int vbr, int cbr,
             double sampleRate, unsigned int numberOfChannels,
-            unsigned int bitsPerSample, const StringPairArray& metadata)
+            int bitsPerSample, const StringPairArray& metadata)
         : AudioFormatWriter (destStream, formatName, sampleRate,
-                             numberOfChannels, bitsPerSample),
+                             numberOfChannels, (unsigned int) bitsPerSample),
           vbrLevel (vbr), cbrBitrate (cbr),
           tempWav (".wav")
     {
